third_party/WebKit/LayoutTests/fast/canvas/canvas-large-pattern-expected.txt

Verifies createPattern using a source image that is a canvas 40k pixels wide.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS imgdata[0] is 0
FAIL imgdata[1] should be 255. Was 0.
PASS imgdata[2] is 0
FAIL imgdata[3] should be 255. Was 0.
PASS imgdata[0] is 0
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[3] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[3] is 0
PASS successfullyParsed is true

TEST COMPLETE
